Make EcsRunLauncher task deduping less strict
Summary:
We observed an issue where subtly different formats of linuxParameters and/or specifying taskRoleArn in differnet formats was causing this logic to incorrectly create a new task definition for each run.

The risk of this is that you might encounter a situation where you try to make a config change and it isn't automatically propagated to runs until you redeploy. Plan is to separately address that by giving you a way to configure a prefix on the run launcher so that you can bump the task definition name when needed.

New test cases

[dagster-aws] Fixed an issue with the EcsRunLauncher where it would sometimes create a new task definition revision for each run if the "task_role_arn" or "execution_role_arn" parameters were specified without the `arn:aws:iam:` prefix.

> Insert changelog entry or delete this section.